A little cleanup and comments
